# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: QA Dashboard API
  version: 3.0.0
paths:
...
  /api/projects/{project_id}/runs:
    get:
      tags:
      - projects
      summary: Get Project Runs
      operationId: get_project_runs_api_projects__project_id__runs_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      - name: active
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Active
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/milestones:
    get:
      tags:
      - projects
      summary: Get Project Milestones
      operationId: get_project_milestones_api_projects__project_id__milestones_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/automation:
    get:
      tags:
      - projects
      summary: Get Project Automation
      operationId: get_project_automation_api_projects__project_id__automation_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/runs/{run_id}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: get_run_api_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: integer
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/runs/{run_id}/results:
    get:
      tags:
      - runs
      summary: Get Run Results
      operationId: get_run_results_api_runs__run_id__results_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: integer
          title: Run Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/notifications/settings:
    get:
      tags:
      - notifications
      summary: Get Settings
      operationId: get_settings_api_notifications_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    put:
      tags:
      - notifications
      summary: Update Settings
      operationId: update_settings_api_notifications_settings_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettingCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/notifications/settings/{project_id}:
    get:
      tags:
      - notifications
      summary: Get Project Settings
      operationId: get_project_settings_api_notifications_settings__project_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/notifications/test:
    post:
      tags:
      - notifications
      summary: Test Notification
      operationId: test_notification_api_notifications_test_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTestPayload'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/feature-flags/admin:
    get:
      tags:
      - feature-flags
      summary: List Admin Flags
      operationId: list_admin_flags_api_feature_flags_admin_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-admin-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - feature-flags
      summary: Create Flag
      operationId: create_flag_api_feature_flags_admin_post
      security:
      - HTTPBearer: []
      parameters:
      - name: x-admin-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/feature-flags/{key}:
    get:
      tags:
      - feature-flags
      summary: Get Flag
      operationId: get_flag_api_feature_flags__key__get
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/feature-flags/admin/{key}:
    put:
      tags:
      - feature-flags
      summary: Update Flag
      operationId: update_flag_api_feature_flags_admin__key__put
      security:
      - HTTPBearer: []
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: x-admin-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - feature-flags
      summary: Delete Flag
      operationId: delete_flag_api_feature_flags_admin__key__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: x-admin-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/integrations/{integration_id}/test:
    post:
      tags:
      - integrations
      summary: Test Integration
      operationId: test_integration_api_integrations__integration_id__test_post
      security:
      - HTTPBearer: []
      parameters:
      - name: integration_id
        in: path
        required: true
        schema:
          type: integer
          title: Integration Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Test Integration Api Integrations  Integration Id  Test
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/testmo-browser/runs/{run_id}/results:
    post:
      tags:
      - testmo-browser
      summary: Add Run Results
      operationId: add_run_results_api_testmo_browser_runs__run_id__results_post
      security:
      - HTTPBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: integer
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Add Run Results Api Testmo Browser Runs  Run Id  Results
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotificationSettingCreate:
      properties:
        project_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Project Id
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        slack_webhook:
          anyOf:
          - type: string
          - type: 'null'
          title: Slack Webhook
        teams_webhook:
          anyOf:
          - type: string
          - type: 'null'
          title: Teams Webhook
        enabled_sla_email:
          type: boolean
          title: Enabled Sla Email
          default: false
        enabled_sla_slack:
          type: boolean
          title: Enabled Sla Slack
          default: false
        enabled_sla_teams:
          type: boolean
          title: Enabled Sla Teams
          default: false
        email_template:
          anyOf:
          - type: string
          - type: 'null'
          title: Email Template
        slack_template:
          anyOf:
          - type: string
          - type: 'null'
          title: Slack Template
        teams_template:
          anyOf:
          - type: string
          - type: 'null'
          title: Teams Template
      type: object
      title: NotificationSettingCreate
    NotificationTestPayload:
      properties:
        channel:
          type: string
          title: Channel
          default: email
        destination:
          anyOf:
          - type: string
          - type: 'null'
          title: Destination
      type: object
      title: NotificationTestPayload
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer