# backend/routes/runs.routes.ts
import express from 'express';
const router = express.Router();
import testmoService from '../services/testmo.service';
import { safeErrorResponse } from '../utils/errorResponse';
import { validateParams, validateQuery, runIdParam, runResultsQuery } from '../validators';

/**
 * Détails d'un run spécifique
 * ISTQB: Test Reporting
 */
router.get('/:runId', validateParams(runIdParam), async (req, res) => {
  try {
    const runId = parseInt(req.params.runId as string);

    const runDetails = await testmoService.getRunDetails(runId);

    res.json({
      success: true,
      data: runDetails,
      timestamp: new Date().toISOString(),
    });
  } catch (error) {
    res.status(500).json(safeErrorResponse(error, `GET /api/runs/${req.params.runId}`));
  }
});

/**
 * Résultats détaillés d'un run
 * API Testmo 2025: Nouveau endpoint
 */
router.get('/:runId/results', validateParams(runIdParam), validateQuery(runResultsQuery), async (req, res) => {
  try {
    const runId = parseInt(req.params.runId as string);
    const { status: statusFilter } = req.validatedQuery as { status?: string };

    const results = await testmoService.getRunResults(runId, statusFilter);

    res.json({
      success: true,
      data: results,
      timestamp: new Date().toISOString(),
    });
  } catch (error) {
    res.status(500).json(safeErrorResponse(error, `GET /api/runs/${req.params.runId}/results`));
  }
});

export default router;